The XAFS measurement is working with getMCAs (through data-link). It looks it is working when I am checking it with a simulator.

+++ Fio.cpp
@@ -421,7 +421,7 @@ void MainWindow::RecordData( void )
         }
         out << buf;
       } else {
-        QVector<int> vals = SFluo->getCountsInROI();           // vals ¤Ï count
+        QVector<unsigned long> vals = SFluo->getCountsInROI(); // vals ¤Ï count
         QVector<double> darks = SFluo->getDarkCountsInROI();   // darks ¤Ï cps 
         for ( int j = 0; j < MaxSSDs; j++ ) {   // 19ch SSD  in ROI
           buf.sprintf(" %9d",
